Improved virtual allocation algo. Added mip zone allocator. Misc

+++ HtLib/System/sys_thread.h
@@ -18,7 +18,7 @@ struct sys_thread
 using PfnSysThreadProc = u32( THREAD_CALLING_CONV * )( void* );
 
 sys_thread  SysCreateThread( u64 stackSize, PfnSysThreadProc ThreadProc, void* pData, const wchar_t* name );
-void        SysThreadSleep( u64 milliSecs );
+void        SysThreadSleep( u32 milliSecs );
 void        SysNameThread( u64 hThread, const wchar_t* name );
 
 #endif //!__HELLTECH_SYS_THREAD_H__
